Add support to install certificate and set web server settings in minimal OS (Docker, Podman, LXC...)
If systemd are rc-service are not available, try to reload the web server configuration using reload command provided by web server. This change is for Apache and Nginx web servers. Lighttpd is not considered here.

Thanks to @megapegabot for detailed report and proposed solution.

This feature request was reported by @megapegabot in bacularis-common repository:

https://github.com/bacularis/bacularis-common/issues/4